+++ object_tracking_cli/processing.py
@@ -9,6 +9,7 @@ from .video_streaming import VideoStream
 from typing import Dict
 
 from .utils.image_utils import resize_with_aspect_ratio
+from .plotting import plot_bboxes, plot_tracking
 
 
 def process_frame(
@@ -41,26 +42,6 @@ def wait_for_next_frame(desired_interval: float, start_time: float) -> float:
     return time.time()
 
 
-def plot_tracking(frame, tracker, tracker_name):
-    font = cv2.FONT_HERSHEY_SIMPLEX
-    font_scale = 0.5
-    thickness = 2
-    color = (0, 0, 255)
-    cv2.putText(frame, tracker_name, (10, 30), font, font_scale, color, thickness)
-    for object_id, (x, y) in tracker.objects.items():
-        text = f"ID {object_id}"
-        cv2.putText(
-            frame,
-            text,
-            (x - 10, y - 10),
-            font,
-            font_scale,
-            color,
-            thickness,
-        )
-        cv2.circle(frame, (x, y), 4, color, -1)
-
-
 def make_class_to_color_and_name(class_id_to_name):
     def hex_to_bgr(hex_color):
         hex_color = hex_color.lstrip("#")
@@ -74,21 +55,6 @@ def make_class_to_color_and_name(class_id_to_name):
         class_id: (colors_bgr[class_id], name)
         for class_id, name in class_id_to_name.items()
     }
-
-
-def plot_bboxes(frame, bboxes, class_to_color_and_name):
-    for x1, y1, x2, y2, class_ in bboxes:
-        color, name = class_to_color_and_name[class_]
-        cv2.putText(
-            frame,
-            name,
-            (x1 - 10, y1 - 10),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            0.5,
-            (0, 255, 0),
-            2,
-        )
-        cv2.rectangle(frame, (x1, y1), (x2, y2), color, 2)
 
 
 def process_video(video_path: str, use_kdtree: bool = False, desired_fps: float = 30.0):
